implemented getAll for BigStore to make weak caching work
