+++ backend/src/main/java/com/bakdata/conquery/io/xodus/stores/BigStore.java
@@ -135,7 +135,9 @@ public class BigStore<KEY, VALUE> implements Store<KEY, VALUE> {
 
 	@Override
 	public Collection<VALUE> getAll() {
-		throw new UnsupportedOperationException();
+		List<VALUE> l = new ArrayList<>();
+		metaStore.forEach(e -> l.add(createValue(e.getKey(), e.getValue())));
+		return l;
 	}
 	
 	private Collection<UUID> writeValue(VALUE value) {
