+++ backend/src/main/java/com/bakdata/conquery/io/xodus/stores/IdentifiableStore.java
@@ -10,12 +10,9 @@ import com.bakdata.conquery.util.functions.ThrowingConsumer;
 import lombok.Setter;
 import lombok.experimental.Accessors;
 
-@Accessors(fluent=true) @Setter
 public class IdentifiableStore<VALUE extends Identifiable<?>> extends KeyIncludingStore<IId<VALUE>, VALUE> {
 
 	private final CentralRegistry centralRegistry;
-	private ThrowingConsumer<VALUE> onAdd;
-	private ThrowingConsumer<VALUE> onRemove;
 	
 	public IdentifiableStore(CentralRegistry centralRegistry, Store<IId<VALUE>, VALUE> store) {
 		this(centralRegistry, store, new SingletonNamespaceCollection(centralRegistry));
